Covid19_AL/requirements.txt
tensorflow~=2.6.0
numpy~=1.19.5
kaggle
opencv-python~=4.5.4.58
pandas~=1.3.4
scikit-learn~=1.0.1
matplotlib~=3.4.3
seaborn
mlflow
Pillow~=8.4.0
streamlit~=1.1.0
deep_utils.git
albumentations~=1.1.0
pyngrok
imutils~=0.5.4
keras~=2.6.0
scipy~=1.7.1